+++ src/test/suite/notification.test.ts
@@ -33,7 +33,7 @@ import {
   PROMO_TYPE,
   SAME_NOTIFICATION_ID,
 } from "./utils/testData";
-import * as asExternal from "../../utils/asExternal";
+import * as asExternalUri from "../../utils/asExternalUri";
 
 type OpenWebviewParams = [
   viewType: string,
@@ -231,13 +231,12 @@ suite("Should poll notifications", () => {
       vscode.WebviewPanel
     > = sinon.spy(vscode.window, "createWebviewPanel");
 
-    const asExternalUriSpy = sinon.spy<(uri: Uri) => Promise<Uri>>(() =>
-      Promise.resolve(Uri.parse(LOCAL_HUB_URL))
-    );
+    const asExternalUriSpy = sinon.stub(asExternalUri, "asExternalUri");
 
-    sinon
-      .fake(asExternal.asExternal)
-      .returned(Promise.resolve(Uri.parse(LOCAL_HUB_URL)));
+    asExternalUriSpy.callsFake(() => {
+      console.log("spy was called");
+      return Promise.resolve(Uri.parse(LOCAL_HUB_URL));
+    });
 
     showInformationMessage.onFirstCall().resolves(AN_OPTION_KEY);
 
@@ -308,12 +307,9 @@ suite("Should poll notifications", () => {
       vscode.WebviewPanel
     > = sinon.spy(vscode.window, "createWebviewPanel");
 
-    const asExternalUriSpy = sinon.spy<(uri: Uri) => Promise<Uri>>(() =>
-      Promise.resolve(Uri.parse(LOCAL_HUB_URL))
-    );
-    sinon
-      .fake(asExternal.asExternal)
-      .returned(Promise.resolve(Uri.parse(LOCAL_HUB_URL)));
+    const asExternalUriSpy = sinon.stub(asExternalUri, "asExternalUri");
+
+    asExternalUriSpy.callsFake(() => Promise.resolve(Uri.parse(LOCAL_HUB_URL)));
 
     showInformationMessage.onFirstCall().resolves(AN_OPTION_KEY);
